Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
openai
pydub
python-dotenv
groq
youtube-transcript-api
pydantic